chore: refactor storage category structure
